+++ test/stylelint.test.js
@@ -1,10 +1,10 @@
 import test from 'ava';
 import config from '../stylelint';
 
-test('should return a object', t => {
+test('should return a object', (t) => {
   t.is(typeof config, 'object');
 });
 
-test('should return a object2', t => {
+test('should return a object2', (t) => {
   t.pass();
 });